requirements.txt
matplotlib==3.3.4
numpy==1.20.1
scipy==1.6.2
opencv-python==4.5.2.54
torch==1.9.0
# torchvision==0.10.0
torchgeometry==0.1.2
scikit-image==0.18.1
Pillow==8.2.0

# python myevaluate.py --model results/mscoco/IHN.pth --dataset mscoco --lev0 --iters_lev0 6 --iters_lev1 0 --gpuid 0 --savemat mscoco.mat --savedict mscoco.npy
